move hookbasedfeatureextractor
